intel/ibc: Use a uint16_t for size in hw_grf_reg_create

+++ src/intel/compiler/ibc.c
@@ -52,7 +52,7 @@ ibc_logical_reg_create(ibc_shader *shader,
 }
 
 ibc_reg *
-ibc_hw_grf_reg_create(ibc_shader *shader, uint8_t size, uint8_t align)
+ibc_hw_grf_reg_create(ibc_shader *shader, uint16_t size, uint8_t align)
 {
    ibc_reg *reg = ibc_reg_create(shader, IBC_REG_FILE_HW_GRF);
